fixed migrations

+++ database/migrations/0001_01_01_060000_create_users_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password')->default(Hash::make('Temp1234'));
-            $table->integer('role_id')->default(3);
+            $table->unsignedBigInteger('role_id')->default(3);
             $table->rememberToken();
             $table->timestamps();
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
